misc refactorings and code cleanup
