+++ src/assets/styles/app/sidebar/export-list.scss
