+++ src/filters/pretty-bytes-filter.js
