+++ src/lib/db/clients/mysql.js
@@ -31,7 +31,6 @@ export default async function (server, database) {
   await driverExecuteQuery(conn, { query: 'select version();' });
 
   return {
-    knexType: 'mysql2',
     supportedFeatures: () => ({ customRoutines: true }),
     wrapIdentifier,
     disconnect: () => disconnect(conn),
