+++ src/lib/export/formats/json.ts
@@ -1,11 +1,11 @@
-import { Export, ExportOptions } from "@/lib/export/export";
+import { Export, ExportOptions } from "@/lib/export";
 import { DBConnection, TableOrView, TableFilter } from '@/lib/db/client'
 import indentString from 'indent-string'
 
 interface OutputOptionsJson {
   prettyprint: boolean
 }
-export default class JsonExporter extends Export {
+export class JsonExporter extends Export {
   readonly format: string = 'json'
 
   constructor(
